+++ packages/commercial-lytenyte-pro/src/index.ts
@@ -5,3 +5,4 @@ export { ColumnManager } from "./column-manager/column-manager.js";
 export { GridBox } from "./grid-box/grid-box.js";
 
 export { activateLicense } from "./license.js";
+export { measureText } from "@1771technologies/lytenyte-shared";