+++ packages/lytenyte-core/src/cells/cell.tsx
@@ -20,11 +20,12 @@ const CellImpl = forwardRef<
   const cx = grid.state;
 
   grid.internal.refreshKey.useValue();
+  const base = grid.state.columnBase.useValue();
 
   const row = cell.row.useValue();
 
   const renderers = cx.cellRenderers.useValue();
-  const providedRenderer = children ?? cell.column.cellRenderer;
+  const providedRenderer = children ?? cell.column.cellRenderer ?? base.cellRenderer;
 
   const Renderer = providedRenderer
     ? typeof providedRenderer === "string"