+++ packages/lytenyte-core/src/row-data-source/use-client-data-source-paginated.ts
@@ -53,6 +53,10 @@ export function makeClientDataSourcePaginated<T>(
   const topData = atom(p.topData ?? []);
   const bottomData = atom(p.bottomData ?? []);
 
+  const dataToSrc$ = atom((g) => {
+    return new Map(g(data).map((c, i) => [c, i]));
+  });
+
   const cache = new Map<number, RowLeaf<T>>();
   const centerNodes = atom((g) => {
     const nodes: RowLeaf<T>[] = [];
@@ -211,6 +215,15 @@ export function makeClientDataSourcePaginated<T>(
     };
 
     return comparator;
+  });
+
+  const idToNode = atom((g) => {
+    const map = new Map<string, TreeNode<RowLeaf<T>>>();
+    traverse(g(tree).root, (node) => {
+      map.set(node.id, node);
+    });
+
+    return map;
   });
 
   const initialized = atom(false);
@@ -372,7 +385,7 @@ export function makeClientDataSourcePaginated<T>(
     );
   };
 
-  const rowById = (id: string) => {
+  const rowById = (id: string): RowNode<T> | null => {
     const pinned = rdsStore.get(pinnedIdMap);
     if (pinned.has(id)) return pinned.get(id)!;
 
@@ -400,35 +413,37 @@ export function makeClientDataSourcePaginated<T>(
 
   const rowUpdate = (updates: Map<string | number, any>) => {
     const grid = rdsStore.get(grid$)!;
-    const t = rdsStore.get(tree);
-
     const d = rdsStore.get(data);
-    for (const [key, data] of updates.entries()) {
-      const rowIndex = typeof key === "number" ? key : rowToIndex(key);
+    const idMap = rdsStore.get(idToNode);
+    const dataToSrc = rdsStore.get(dataToSrc$);
 
-      const row = rowByIndex(rowIndex);
-      if (!row || !grid) {
-        console.error(`Failed to find the row at index ${rowIndex} which is being updated.`);
+    for (const [key, next] of updates.entries()) {
+      const row = typeof key === "string" ? rowById(key) : rowByIndex(key);
+      const treeNode = typeof key === "string" ? idMap.get(key) : null;
+
+      if ((!row && !treeNode) || !grid) {
+        console.error(`Failed to find the row with identifier ${key} which is being updated.`);
         continue;
       }
 
-      if (row.kind === "branch") {
-        (row as any).data = data;
+      if (row?.kind === "branch") {
+        (row as any).data = next;
       } else {
-        const source = t.idToSourceIndex.get(row.id);
+        const data = row?.kind === "leaf" ? row.data : treeNode?.data.data;
+
+        const source = dataToSrc.get(data as T);
         if (source == null) {
-          console.error(`Failed to find the row at index ${rowIndex} which is being updated.`);
+          console.error(`Failed to find the row with identifier ${key} which is being updated.`);
           continue;
         }
 
-        d[source] = data;
+        d[source] = next as any;
       }
-
-      grid.state.rowDataStore.rowInvalidateIndex(rowIndex);
     }
 
     rdsStore.set(data, [...d]);
     rdsStore.set(snapshot, (prev) => prev + 1);
+    grid.state.rowDataStore.rowClearCache();
   };
 
   const rowToIndex = (rowId: string) => {
@@ -481,6 +496,12 @@ export function makeClientDataSourcePaginated<T>(
       rowByIndex,
       rowAllChildIds,
       rowUpdate,
+
+      rowSetCenterData: (d: any[]) => {
+        rdsStore.set(data, d);
+        const grid = rdsStore.get(grid$);
+        grid?.state.rowDataStore.rowClearCache();
+      },
       rowAdd: (newRows, place = "end") => {
         rdsStore.set(data, (prev) => {
           if (!newRows.length) return prev;
@@ -633,12 +654,12 @@ export function useClientRowDataSourcePaginated<T>(p: ClientRowDataSourceParams<
   const da = dataAtomRef.current;
   if (p.reflectData) {
     // Need to queue the microtask since it we cannot update state during render.
-    if (p.data !== da.center.get()) queueMicrotask(() => da.center.set(p.data));
+    if (p.data !== da.center.get()) queueMicrotask(() => ds.current.rowSetCenterData(p.data));
     if (!equal(p.topData ?? [], da.top.get())) {
-      queueMicrotask(() => da.top.set(p.topData ?? []));
+      queueMicrotask(() => ds.current.rowSetTopData(p.topData ?? []));
     }
     if (!equal(p.bottomData ?? [], da.bottom.get()))
-      queueMicrotask(() => da.bottom.set(p.bottomData ?? []));
+      queueMicrotask(() => ds.current.rowSetBotData(p.bottomData ?? []));
   }
 
   return ds.current;