+++ packages/fork-focus-trap/CHANGELOG.md
@@ -1,5 +1,59 @@
 # @1771technologies/lytenyte-focus-trap
 
+## 1.0.0-beta.10
+
+### Patch Changes
+
+- Version
+- Updated dependencies
+  - @1771technologies/lytenyte-focus@1.0.0-beta.10
+  - @1771technologies/lytenyte-dom-utils@1.0.0-beta.10
+
+## 1.0.0-beta.9
+
+### Patch Changes
+
+- Version
+- Updated dependencies
+  - @1771technologies/lytenyte-focus@1.0.0-beta.9
+  - @1771technologies/lytenyte-dom-utils@1.0.0-beta.9
+
+## 1.0.0-beta.8
+
+### Patch Changes
+
+- Version
+- Updated dependencies
+  - @1771technologies/lytenyte-focus@1.0.0-beta.8
+  - @1771technologies/lytenyte-dom-utils@1.0.0-beta.8
+
+## 1.0.0-beta.7
+
+### Patch Changes
+
+- Fixes
+- Updated dependencies
+  - @1771technologies/lytenyte-focus@1.0.0-beta.7
+  - @1771technologies/lytenyte-dom-utils@1.0.0-beta.7
+
+## 1.0.0-beta.6
+
+### Patch Changes
+
+- Version
+- Updated dependencies
+  - @1771technologies/lytenyte-focus@1.0.0-beta.6
+  - @1771technologies/lytenyte-dom-utils@1.0.0-beta.6
+
+## 1.0.0-beta.5
+
+### Patch Changes
+
+- Version
+- Updated dependencies
+  - @1771technologies/lytenyte-focus@1.0.0-beta.5
+  - @1771technologies/lytenyte-dom-utils@1.0.0-beta.5
+
 ## 1.0.0-beta.4
 
 ### Patch Changes