+++ packages/fork-object-equals/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @1771technologies/lytenyte-object-equals
 
+## 1.0.0-beta.10
+
+### Patch Changes
+
+- Version
+
+## 1.0.0-beta.9
+
+### Patch Changes
+
+- Version
+
+## 1.0.0-beta.8
+
+### Patch Changes
+
+- Version
+
+## 1.0.0-beta.7
+
+### Patch Changes
+
+- Fixes
+
+## 1.0.0-beta.6
+
+### Patch Changes
+
+- Version
+
+## 1.0.0-beta.5
+
+### Patch Changes
+
+- Version
+
 ## 1.0.0-beta.4
 
 ### Patch Changes